+++ src/commands/EXPECT/expectCommandParser.ts
@@ -4,7 +4,7 @@ import { parseNumber } from '../_common/parseNumber';
 import type { CommandParser, CommandParserInput } from '../_common/types/CommandParser';
 import type { ExpectAmountCommand } from './ExpectAmountCommand';
 import type { ExpectCommand } from './ExpectCommand';
-import { ExpectFormatCommand } from './ExpectFormatCommand';
+import type { ExpectFormatCommand } from './ExpectFormatCommand';
 
 /**
  * Parses the expect command