+++ src/types/PipelineJson/LlmTemplateJson.ts
@@ -5,11 +5,11 @@ import type { PromptTemplateJsonCommon } from './PromptTemplateJsonCommon';
  * Template for prompt to LLM
  */
 export type LlmTemplateJson = PromptTemplateJsonCommon & {
-    readonly executionType: 'PROMPT_TEMPLATE';
+    readonly blockType: 'PROMPT_TEMPLATE';
 
     /**
      * Requirements for the model
-     * - This is required only for executionType PROMPT_TEMPLATE
+     * - This is required only for blockType PROMPT_TEMPLATE
      */
     readonly modelRequirements: ModelRequirements; // <- TODO: !!! Make Partial<ModelRequirements> and optional;
 };