+++ src/types/PipelineJson/PromptDialogJson.ts
@@ -4,5 +4,5 @@ import type { PromptTemplateJsonCommon } from './PromptTemplateJsonCommon';
  * Template for prompt to user
  */
 export interface PromptDialogJson extends PromptTemplateJsonCommon {
-    readonly executionType: 'PROMPT_DIALOG';
+    readonly blockType: 'PROMPT_DIALOG';
 }