+++ src/types/PipelineJson/ScriptJson.ts
@@ -5,11 +5,11 @@ import type { PromptTemplateJsonCommon } from './PromptTemplateJsonCommon';
  * Template for script execution
  */
 export interface ScriptJson extends PromptTemplateJsonCommon {
-    readonly executionType: 'SCRIPT';
+    readonly blockType: 'SCRIPT';
 
     /**
      * Language of the script
-     * - This is required only for executionType SCRIPT
+     * - This is required only for blockType SCRIPT
      *
      */
     readonly contentLanguage?: ScriptLanguage;