+++ src/types/PipelineJson/SimpleTemplateJson.ts
@@ -4,5 +4,5 @@ import type { PromptTemplateJsonCommon } from './PromptTemplateJsonCommon';
  * Template for simple concatenation of strings
  */
 export interface SimpleTemplateJson extends PromptTemplateJsonCommon {
-    readonly executionType: 'SIMPLE_TEMPLATE';
+    readonly blockType: 'SIMPLE_TEMPLATE';
 }